Restructure from lib to client/server directory structure
