+++ test/index.js
@@ -1,6 +1,7 @@
 require('tape').onFailure(() => process.exit(1));
 
-require('./utils/validateAlphaNumericDashDot');
+require('../utils/validateAlphaNumericDashDot');
 require('./client');
+require('./notify');
 require('./locking');
 require('./integration');
