+++ utils/validateAlphaNumericDashDot.js
