+++ utils/validateQueryOptions.js
